Move test_user_get to helevents app
Also add test for user list endpoint.

Refs LINK-1122
