+++ events/tests/conftest.py
@@ -1,12 +1,10 @@
 from datetime import datetime, timedelta
 
 # 3rd party
-from django.contrib.auth import get_user_model
 from django.contrib.gis.geos import MultiPolygon, Point, Polygon
 
 # django
 from django.utils import timezone
-from django_orghierarchy.models import Organization
 from munigeo.models import (
     AdministrativeDivision,
     AdministrativeDivisionGeometry,
@@ -14,13 +12,11 @@ from munigeo.models import (
     Municipality,
 )
 from parler.utils.context import switch_language
-from rest_framework.test import APIClient
 
 from events.api import KeywordSerializer, LanguageSerializer, PlaceSerializer
 
 # events
 from events.models import (
-    DataSource,
     Event,
     Keyword,
     KeywordLabel,
@@ -42,8 +38,6 @@ TEXT_EN = "testing"
 URL = "http://localhost"
 DATETIME = (timezone.now() + timedelta(days=1)).isoformat().replace("+00:00", "Z")
 
-OTHER_DATA_SOURCE_ID = "testotherdatasourceid"
-
 
 @pytest.fixture
 def kw_name():
@@ -58,108 +52,6 @@ def kw_name_2():
 @pytest.fixture
 def kw_name_3():
     return "känd_nyckelord"
-
-
-@pytest.fixture
-def api_client():
-    return APIClient()
-
-
-@pytest.fixture()
-def user_api_client(user):
-    api_client = APIClient()
-    api_client.force_authenticate(user)
-    return api_client
-
-
-@pytest.mark.django_db
-@pytest.fixture
-def data_source():
-    return DataSource.objects.create(
-        id=settings.SYSTEM_DATA_SOURCE_ID,
-        api_key="test_api_key",
-        user_editable_resources=True,
-        user_editable_organizations=True,
-    )
-
-
-@pytest.mark.django_db
-@pytest.fixture
-def other_data_source():
-    return DataSource.objects.create(id=OTHER_DATA_SOURCE_ID, api_key="test_api_key2")
-
-
-@pytest.mark.django_db
-@pytest.fixture
-def user():
-    return get_user_model().objects.create(
-        username="test_user", first_name="Cem", last_name="Kaner", email="cem@kaner.com"
-    )
-
-
-@pytest.mark.django_db
-@pytest.fixture
-def user2():
-    return get_user_model().objects.create(
-        username="test_user2",
-        first_name="Brendan",
-        last_name="Neutra",
-        email="brendan@neutra.com",
-    )
-
-
-@pytest.mark.django_db
-@pytest.fixture
-def super_user():
-    return get_user_model().objects.create(
-        username="super_user",
-        first_name="Super",
-        last_name="Man",
-        email="super@user.com",
-        is_superuser=True,
-    )
-
-
-@pytest.mark.django_db
-@pytest.fixture
-def organization(data_source, user):
-    org, created = Organization.objects.get_or_create(
-        id=data_source.id + ":test_organization",
-        origin_id="test_organization",
-        name="test_organization",
-        data_source=data_source,
-    )
-    org.admin_users.add(user)
-    org.save()
-    return org
-
-
-@pytest.mark.django_db
-@pytest.fixture
-def organization2(other_data_source, user2):
-    org, created = Organization.objects.get_or_create(
-        id=other_data_source.id + ":test_organization2",
-        origin_id="test_organization2",
-        name="test_organization2",
-        data_source=other_data_source,
-    )
-    org.admin_users.add(user2)
-    org.save()
-    return org
-
-
-@pytest.mark.django_db
-@pytest.fixture
-def organization3(other_data_source, user2):
-    org, created = Organization.objects.get_or_create(
-        id=other_data_source.id + ":test_organization3",
-        origin_id="test_organization3",
-        name="test_organization3",
-        data_source=other_data_source,
-    )
-    org.admin_users.add(user2)
-    org.save()
-    return org
 
 
 @pytest.mark.django_db
